extract resolveinput
